requirements.txt
appdirs==1.4.4
asgiref==3.3.0
autopep8==1.5.4
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-geojson==3.1.0
django-leaflet==0.27.1
django-pwa==1.0.10
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-gis==0.16
filelock==3.0.12
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2020.4
requests==2.25.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
virtualenv==20.2.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
